requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
Jinja2==3.1.2
PyYAML==6.0.1

apache-airflow-providers-amazon==5.0.0
apache-airflow-providers-snowflake==3.2.0
astro-sdk-python==1.1.0